Create deliverable concern
